Implement result and remediation forwarding
This commit identifies a basic seam to implement result and remediation
forwarding. We're using a factory to return a forwarding implementation
that defaults to a noop forwarder. If a scan setting uses `Debug: true`,
the factory will return a log forwarder.

Subsequent patches will include support for a gRPC forwarding
implementation.