Ensure ScanSettingBindings use the default ScanSetting by default
By default, ScanSettingBindings don't need to reference a ScanSetting
using the SettingsRef. If you create a binding without a reference to a
scan setting, the scan will run anyway but doesn't take on any of the
default behaviors of the `default` ScanSetting.

This commit fixes this behavior by updating the default for the
ScanSettingBinding settingRef to the `default` ScanSetting.

Fixes Jira: https://issues.redhat.com/browse/OCPBUGS-3420