ssb: Error if attempting to use a non-default pool but the pool variables are not set
Since we switched to the new way of remediating kubeletConfigs, setups
that use a non-default pool name require that the pool is set in
variables in a tailored profile. Otherwise, some remediations might be
set cluster-wide which can cause issues, e.g. if the platform-level remediation that
sets protectKernelDefaults is set cluster-wide, but the node level
remediation that sets the required sysctls is only set pool-wide.

Protect the user againts shooting themselves in the foot by refusing to
process this kind of SSB CRs.

Resolves: OCPBUGS-3864