OCPBUGS-6710: remediations: Normalize remediations' annotations to avoid needlesly marking remediations as Outdated
Because the remediation contents we Get() from the cluster after they've
been applied contain an empty Annotations map, but the remediations we
parse from content do not, the comparison always yielded false and the
remediatins have been set to Outdated as a result.

This patch adds an empty annotations map to remediations before
comparing them, so that we compare their normalized version.