Add Makefile target for generating unit test coverage reports
While end-to-end coverage is important, it's good to have an idea of
what parts of the code aren't being exercised with unit tests. Good unit
tests can provide efficent coverage without expensive infrastructure or
wait times.

This commit makes it easier for developers to see test coverage by
adding an explicity target for it.