Add makefile targets to run serials and parallel e2e tests
The e2e test framework keeps track of tests that are safe to run
concurrently and which must be run serially.

For example, tests that focus on scanning and don't interfere with each
other are safe to run in parallel. Tests that apply remediations to the
cluster impact other tests that assert specific checks or outcomes from
a scan.

We typically lump all these tests into a single run using `make e2e` and
ensure the concurrent tests are run first, followed by the serial tests.

By adding separate makefile targets, we're making it easier for
contributors to run a group of tests, instead of the whole suite,
shortening the feedback loop. We can also use this in CI and create
separate jobs, running parallel tests and serial tests on different
clusters, speeding up CI reporting time. It's also easier to see, at a
glance, if a particular set of tests is failing consistently.

This change also adds a layer of abstraction between contributors and
the internal details of the tests, giving us more flexibility to evolve
the tests and reorganize them as we see fit without as much disruption
to contributors.