Add cleanup logging after tests run
We've noticed that some of the tests aren't cleaning up the way we
expect. Specifically, some of the compliance scans are still laying
around even when we exit the context Cleanup() function, which is
designed to execute a bunch of callbacks to cleanup test resources.

This probably isn't much of a problem in CI since the cluster gets torn
down after the tests run, but it impacts contributor's ability to re-use
the same environment for multiple e2e runs. Ideally, we want to run a
single test in isolation and have it cleanup after itself so we can
continue using the same environment.

This commit adds basic logging in Cleanup() that logs how many scans are
left over before we tear down the environment.