Refactor e2e test cleanup
The e2e tests use a nested test structure, where TestE2E is the
grandparent test, TestE2E/Serial_tests and TestE2E/Parallel_tests are
the parent tests, and each test case is technically a grandchild of
TestE2E.

We use this structure so that TestE2E can actually setup shared
resources for the test (like deploying and configuring the operator).

All the setup functionality used the same .Create() method available
through the test framework client, which automatically defers cleanup of
those resources by executing the deferred calls in a FILO order (stack).

However, as we've seen in various test runs, it's possible for CRDs to
get cleaned up before all resources have been deleted. This orphans the
resources, making them harder to cleanup, leaving the operator resources
and it's namespace in a terminating state.

While this probably isn't a huge deal for CI, which throws away the
environment after tesing, it's problematic when you're developing e2e
tests locally and re-using the same cluster.

This commit adds a specific layer in the test runner that is guaranteed
to run after all tests execute, and cleans up resources in a specific
order so that we don't orphan resources created during the test runs. It
also adds a separate Create function to the framework client that
creates resources without deferring cleanup.

Fixes #231