Simplify skipCleanupOnError test flag
The flag skipCleanupOnError defaulted to false, meaning it would cleanup
test resources in the event of a failure. In the Makefile target for e2e
tests, we override -skipCleanupOnError with true, so we keep resources
when a test fails.

Depending on if you're running the tests using `make e2e` or directly
using `go test` you'd get different behavior.

This commit renames the variable to be more affimative by removing
`skip` from the name (skipCleanupOnError actually means
cleanupOnError=false). It also converges the default in the test runner
to be the same as the default in the Makefile, allowing for consistent
behavior regardless of how you invoke the tests. It also adds more help
text to clarify the behavior.