Refactor e2e test setup
We recently made a change to ensure we clean up resources properly after
we perform e2e tests. Specifically to avoid race conditions that leaves
the compliance operator resources in a bricked state, even when the
tests pass.

This commit uses a similar pattern for test setup by leveraging the
`setUp` method before calling `m.Run()` to make sure the cluster is
ready before we run the actual tests.

One advantage is that we're simplifying the logic to a single layer
before the tests run.

Further changes will make it easier to decouple the framework, context,
and e2eutils into a single set of utilities we can share across
repositories and test suites.