Add parallel test module
Now that the framework is mostly decoupled from the tests, we can reuse
it for separate test modules for parallel and serial tests.

The benefit of this is that we can flatten the tests. Which aid
readability, reduces the ammount of variables we need to pass into tests
regardless if they're used, allows us to use native go tools for test
discovery, and gives use the opportunity to remove the custom test
runner implementation we have to handle parallel and serial tests.