Fix upstream repository URL
We moved the code from openshift/compliance-operator to
ComplianceAsCode/compliance-operator, but we did't update the manifest
references. This means the bundle image would still point to the origin
repository, which isn't what we want.

We attempted to fix this in PR #230, but we only modified the bundle
manifest, which is generated from the manifest base. This commit updates
the manifest base so that it uses the correct URL, too.