Fix TestSingleTailoredPlatformScanSucceeds by using supported rule
The TestSingleTailoredPlatformScanSucceeds creates a simple tailored
profile with a single rule, binds it to a scan setting, then asserts the
scan was compliant.

This recently started failing because it was using the
scheduler_no_bind_address rule, which is not applicable to newer
versions of OpenShift:

  https://github.com/ComplianceAsCode/content/commit/778fae292ca76a16b3a191b0a1f75d9241e16dd0

This resulted in the scan result coming back as not applicable instead
of compliant.

This commit uses a different rule for the tailored profile that's still
supported on newer versions of OpenShift so our CI works again.

Fixes #276