OCPBUGS-4409: Support converting HTML links to Markdown
The operator uses a handy library for converting HTML-like text to
markdown. This is nice for content text that informs users about what to
do with specific findings because it's formatted (e.g., adding lines
around formatted code snippet).

Unfortunately, the parser didn't account for links, which contain
multiple XML elements. The original `<a></a>` tag was preserved but the
actual link reference in `href=example.com` was dicarded.

This results in none of the links we put in content getting rendered in
the actual CRs where users can read and use them.

This commit adds some special handling for XML start elements to detect
cases where we're dealing with links and ensures they get propogated
into the final Markdown representation.