Run 'make bundle' to commit bundle manifest changes
In cac/content testing, we are now using nightly operator images, which
are built after every commit to cac/compliance-operator. However,
bundle manifests are only refreshed on release, which means that some
code changes like rbac modifications are not reflected.

This causes CI failures in cac/content unless we release.