Allow setting the OSCAP_DEBUG_LEVEL to tune OpenScap debug level
In some cases it's handy to enable the DEVEL debug level for OpenScap,
e.g. to diagnose issues in the scanner itself.

In order to keep the semantics of the current `debug` attribute of the
scans, but still allow setting the very verbose debug level, let's
take the value of the `OSCAP_DEBUG_LEVEL` environment variable into
account.

The variable can be set either directly in the operator deployment or
through OLM Subscription.