Adjust CO reconcile loop setup and unit tests for the new controller-manager version
In order to use the latest release of controller-manager we need to use
the new style of setting up controller loops that uses
ctrl.NewControllerManagedBy() and then the builder pattern instead of
adding c.Watch for each resource.

The tests had to be adjusted, too, to specifically name resources that
can update its status.