OLM: Provide an OCP-specific bundle kustomize
Commit 03ef0398dd26d85737bfdbf6fe14f37b116abbb6 removed the OCP-specific
nodeSelector and tolerations from the default deployment at
`config/manager/deployment.yaml`. This would mean that the bundle
wouldn't have them at all, which is fine for generic Kube deployments,
but we still want to have them on OCP.

This patch introduces an OCP specific bundle kustomization that applies
a patch to the manager config that patches back the nodeSelector and the
tolerations.