Update UBI image to keep pace with GLIBC version
If you build the Compliance Operator from source and deploy it into a
cluster using `make deploy-local`, you'll see that the operator fails to
start because of a GLIBC error:

  compliance-operator: /lib64/libc.so.6: version `GLIBC_2.32' not found (required by compliance-operator)
  compliance-operator: /lib64/libc.so.6: version `GLIBC_2.34' not found (required by compliance-operator)

This is because the container image for golang:1.20 is referencing a
newer version fo GLIBC than what's available in the ubi8 minimal image.

This commit updates the ubi image so that we keep up with the GLIBC
version being referenced in newer golang versions.

Fixes #372