CMP-2176: Implement `replaces` in bundle CSV
The `replaces` attribute of an operator's CSV helps OLM build
relationships between operators, making it easier for users to install
older versions of an operator (or know which version supercedes the
version they're using).

This commit reintroduces the `replaces` attribute, since it was being
deleted before as a side-effect of an operator-sdk update. It should be
safe to include again.