Check for truthy values with new annotations
This migration/feature adds new annotations that help users manage their
TailoredProfiles. For example, they can set an annotation that
automatically prune outdated rule references from the TailoredProfile.

Previously, we were only checking the presence of these annotations on
the TailoredProfile. This meant users should invoke the behavior with an
empty string (""), or a truthy value ("true"), or even a falsy ("false)
value.

This commit explicitly checks the value of the annotation and makes sure
it's set to true before invoking the behavior.