Remove deprecation code for handling rules and variables
We have a pretty good understanding of how to migrate rules from one
type (Node) to another (Platform), but initially we included some
thoughts on how to handle deprecated rules and variables.

However, up to this point, we don't have a way to flag deprecated rules
or variables in the content, which is what's needed to annotate them
accordingly.

Since this code is involving a separate case (deprecation), and the
current migration e2e tests pass without it, let's remove it so we can
come up with a dedicated approach for it later.

This cuts down on the overall patch size, making it more digestable for
reviewers, especially since it wasn't fully tested.