Use git describe for finding operator version
We need to be able to set the `replaces` attribute of the CSV file so
that OLM can build a relationship between dependencies, which is
important for disconnected installations and upgrades.

Previously, we were using the version in the CSV file, which worked but
it would change within the `make prepare-release` target depending on
when the command was invoked.

This commit updates the version script to use `git describe` instead,
which makes finding the current version more robust since tagging isn't
done until a subsequent step of the release process.