Automatically detect if operator is running on ROSA
Made change to be able to detect what managed cluster we are on using
ClusterClaim and added ROSA platform type.

This makes it easier for users to run the operator on ROSA, without
having to specify the platform. Additional patches can use this to by
smarter about loading in specific resources on ROSA, like profiles and
scan configs.