CMP-2526: Disable automatic remediation for ROSA HCP environments
Since the Compliance Operator is only going to support running on HCP environments,
where end users don't have access to the master nodes (or control plane),
it doesn't make sense to give them a scan setting with auto remediation functionality applied.

For this case, the Compliance Operator should detect if it is running on ROSA HCP and adjust,
or just not create the `default-auto-apply` scan settings.