CMP-1096: Add must-gather image to relatedImages
Even though the must-gather plugin image isn't necessary for running the
operator, is helps users collect information when reporting issues.

By including the image reference in the `relatedImages` of the CSV, it's
more discoverable and users can query it out of the CSV directly,
instead of having to hunt through documentation and container
registries.