CMP-2693: Use CLI image for base image in must-gather
The original approach for building a must-gather image specifically for
Compliance Operator usage grabbed the latest stock must-gather image
(the one for collecting everything in an ordinary deployment), and then
pushing the original scripts out of the way and replacing them with
Compliance Operator specific scripts to collect the information we
wanted.

While this worked, we can simplify the image dependency by just relying
on the CLI image directly, since that's what the upstream must-gather
image does, then just wire up the entry point to the collection scripts
we already have, following the same pattern that the upstream
must-gather image uses.

This commit also updates the `Dockerfile` name to include `.ocp` suffix,
since we're relying on a Red Hat registry to pull the CLI image.