CMP-2196: Update cluster role permissions for ingresscontrollers
We have a rule in the compliance content that will check the
ingresscontroller's TLS ciphers. It also comes with a remediation to
update the ciphers so they're either Recommended or Secure. However,
the remediation errors out when it's loaded after a scan because the
Compliance Operator doesn't have the necessary permissions to get the
ingresscontrollers, or apply the remediation.

This commit updates the Compliance Operator's cluster role so that it
has the ability to update the TLS cipher configuration for
ingresscontrollers.