OCPBUGS-37697: Remove newline in metrics doc literal
We have an example in our documentation that tells users how to fetch
metrics using the command line. But, the command has a newline in the
middle of the authorization header, and if you copy/paste it from the
document, like most readers would, it returns an HTTP 404.

Remove the newline so the command works when you copy/paste it from the
documentation, which is the intention.