Remove multi-arch support from must-gather image
We're struggling to find a multi-arch image we can use for building the
must-gather image. This causes the GitHub action we have wired up to
build the image automatically to fail.

This commit removes power and z support for now so we can get some
builds produces that incorporate the new must-gather tooling.