Update bundle manifests
The github action we have for making sure we update the bundle in PRs
broke, causing some changes to slip into the operator without updating
the bundle manifests.

This commit updates the bundle by running `make verify-bundle` then
committing the changes.

The manifests changes reflect updates we made to the operator so that it
has the necessary permissions for ingresscontrollers and
kubedeschedulers.