Update bundle annotations to include must-gather image
Since we have a must-gather image, we should advertise that information
through the bundle. We typically do this using `RELATED_IMAGES`, which
the operator uses to find container images it needs to do things like
profile parsing or scanning.

While we can use `RELATED_IMAGES` for the must-gather image, the
Compliance Operator doesn't use the must-gather image directly. Instead,
we can put it in the annotation of the bundle CSV, so it's still
discoverable to users, while not polluting the operator environment with
variables it doesn't need.